Update maxTweeter.c
okay i put the checks in except the comma one. besides that i added a function to get the actual line count so we have a variable we can use to allocate things. the prompt says the file can have at most 20000 lines so we will have files with less than that.